change the label for subsection "Predicates"
The HTML id label of the section "Predicates" clashes with that of the subsection "Predicates" under "Language overview". When clicking on the navigation menu item "Predicates", it jumps to the subsection "Predicates" instead of the "Predicates" on the top level. This commit fixes the bug by manually specifying the id of subsection "Predicates".